Add files via upload
The number of well ordered permutations will be displayed in the console. In the program will output the permutations in a text document. This document can be found in the source folder for the program. I would recommend not trying this with strings longer than 6 digits. Past this point the file size becomes exceedingly huge (2gb - 3gb), and you may have trouble opening the .txt file.